pyproject.toml
[tool.poetry]
name = "getdatafaster"
version = "0.1.0"
description = ""
authors = ["Luciano Filho <lvgalvaofilho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
faker = "^20.1.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
